chore(dbaas): migrate to new database package
Move DBaaS related code to new `database` service package.
