+++ upcloud/resource_upcloud_managed_database_test.go
@@ -1,10 +1,8 @@
 package upcloud
 
 import (
-	"context"
 	"io/ioutil"
 	"testing"
-	"time"
 
 	"github.com/UpCloudLtd/upcloud-go-api/v4/upcloud"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
@@ -92,52 +90,4 @@ func TestAccUpcloudManagedDatabase(t *testing.T) {
 			},
 		},
 	})
-}
-
-// TODO move the tests below to db utils test file, once the managed database resources are moved under their own package
-func TestIsManagedDatabaseFullyCreated(t *testing.T) {
-	db := &upcloud.ManagedDatabase{
-		Backups: make([]upcloud.ManagedDatabaseBackup, 0),
-		State:   upcloud.ManagedDatabaseStatePoweroff,
-		Users:   make([]upcloud.ManagedDatabaseUser, 0),
-	}
-	if isManagedDatabaseFullyCreated(db) {
-		t.Errorf("isManagedDatabaseFullyCreated failed want false got true %+v", db)
-	}
-
-	db.State = upcloud.ManagedDatabaseStateRunning
-	db.Backups = append(db.Backups, upcloud.ManagedDatabaseBackup{})
-	if isManagedDatabaseFullyCreated(db) {
-		t.Errorf("isManagedDatabaseFullyCreated failed want false got true %+v", db)
-	}
-
-	db.Users = append(db.Users, upcloud.ManagedDatabaseUser{})
-	db.Backups = make([]upcloud.ManagedDatabaseBackup, 0)
-	if isManagedDatabaseFullyCreated(db) {
-		t.Errorf("isManagedDatabaseFullyCreated failed want false got true %+v", db)
-	}
-
-	db.Backups = append(db.Backups, upcloud.ManagedDatabaseBackup{})
-	if !isManagedDatabaseFullyCreated(db) {
-		t.Errorf("isManagedDatabaseFullyCreated failed want true got false %+v", db)
-	}
-}
-
-func TestWaitServiceNameToPropagate(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*2)
-	defer cancel()
-	name := "upcloud.com"
-	if err := waitServiceNameToPropagate(ctx, name); err != nil {
-		t.Errorf("waitServiceNameToPropagate failed %+v", err)
-	}
-}
-
-func TestWaitServiceNameToPropagateContextTimeout(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 1)
-	defer cancel()
-	name := "upcloud.com"
-	if err := waitServiceNameToPropagate(ctx, name); err == nil {
-		d, _ := ctx.Deadline()
-		t.Errorf("waitServiceNameToPropagate failed didn't timeout before deadline %s", d.Format(time.RFC3339))
-	}
 }
