+++ upcloud/resource_upcloud_objectstorage.go
@@ -189,7 +189,7 @@ func resourceObjectStorageRead(ctx context.Context, d *schema.ResourceData, m in
 	if err != nil {
 		if svcErr, ok := err.(*upcloud.Error); ok && svcErr.ErrorCode == upcloudObjectStorageNotFoundErrorCode {
 			var diags diag.Diagnostics
-			diags = append(diags, diagBindingRemovedWarningFromUpcloudErr(svcErr, d.Get("name").(string)))
+			diags = append(diags, utils.DiagBindingRemovedWarningFromUpcloudErr(svcErr, d.Get("name").(string)))
 			d.SetId("")
 			return diags
 		}
