+++ upcloud/resource_upcloud_server.go
@@ -455,7 +455,7 @@ func resourceUpCloudServerRead(ctx context.Context, d *schema.ResourceData, meta
 	server, err := client.GetServerDetails(r)
 	if err != nil {
 		if svcErr, ok := err.(*upcloud.Error); ok && svcErr.ErrorCode == upcloudServerNotFoundErrorCode {
-			diags = append(diags, diagBindingRemovedWarningFromUpcloudErr(svcErr, d.Get("hostname").(string)))
+			diags = append(diags, utils.DiagBindingRemovedWarningFromUpcloudErr(svcErr, d.Get("hostname").(string)))
 			d.SetId("")
 			return diags
 		}
